fix: fix and refactor codes to make it work as intended

+++ src/code/ca/bcit/comp2522/project/Score.java
@@ -79,10 +79,10 @@ public class Score
      *
      * @return the average score for this session
      */
-    public int getAverageScore()
+    public double getAverageScore()
     {
-        final int averageScore;
-        averageScore = this.getScore() / numGamesPlayed;
+        final double averageScore;
+        averageScore = (double) this.getScore() / numGamesPlayed;
         return averageScore;
     }
